JM- fix and refactored api/test routes
