--- routes/api/mockData/mockData.json
@@ -1,32 +1,0 @@
-[
-    {
-        "firstName": "Edwin",
-        "lastName": "Holt",
-        "email": "test@test.com",
-        "password": "12345"
-    },
-    {
-        "firstName": "Clayton",
-        "lastName": "Jones",
-        "email": "up@cefava.cg",
-        "password": "cjoweir"
-    },
-    {
-        "firstName": "Max",
-        "lastName": "Green",
-        "email": "tuidmut@acouhujo.gm",
-        "password": "weproit947fu"
-    },
-    {
-        "firstName": "Lilly",
-        "lastName": "Obrien",
-        "email": "ovutazip@ad.gm",
-        "password": "f049vjaso"
-    },
-    {
-        "firstName": "Augusta",
-        "lastName": "Schneider",
-        "email":    "ti@wi.gy",
-        "password": "perovojer90"
-    }
-]
